+++ crates/mimic/src/db/executor/delete.rs
@@ -74,10 +74,10 @@ impl DeleteExecutor {
     }
 
     // filter_eq
-    pub fn filter_eq<E: EntityKind, F: Into<String>, V: Into<Value>>(
+    pub fn filter_eq<E: EntityKind>(
         self,
-        field: F,
-        value: V,
+        field: &str,
+        value: impl Into<Value>,
     ) -> Result<Vec<Key>, MimicError> {
         self.execute::<E>(DeleteQuery::new().filter_eq(field, value))
     }